+++ db/migrate/20220613000000_extend_tables7.rb
@@ -1,0 +1,5 @@
+class ExtendTables7 < ActiveRecord::Migration[6.0]
+  def change
+    add_column :tables, :yn_initialize_with_flashback,  :string, limit: 1, null: false, default: 'Y', comment: 'Should flashback query used to init only records before create trigger SCN'
+  end
+end