+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2022_04_14_000000) do
+ActiveRecord::Schema.define(version: 2022_06_13_000000) do
 
   create_table "activity_logs", force: :cascade do |t|
     t.integer "user_id", precision: 38, null: false, comment: "Reference to user"
@@ -127,6 +127,7 @@ ActiveRecord::Schema.define(version: 2022_04_14_000000) do
     t.string "yn_initialization", limit: 1, default: "N", null: false, comment: "Should current content of table be transferred to Kafka as insert events at next trigger generation?"
     t.string "initialization_filter", limit: 4000, comment: "SQL filter expression to filter current content of table before transferred to Kafka as insert events at next trigger generation"
     t.string "initialization_order_by", limit: 4000, comment: "Optionally sort current content of table before transferred to Kafka as insert events at next trigger generation"
+    t.string "yn_initialize_with_flashback", limit: 1, default: "Y", null: false, comment: "Should flashback query used to init only records before create trigger SCN"
     t.index ["schema_id", "name"], name: "ix_tables_schema_name", unique: true
     t.index ["schema_id"], name: "index_tables_on_schema_id"
   end